Prijs Int naar Float

+++ app/model/Item.php
@@ -4,7 +4,7 @@ class Item{
     public int $itemId;
     public string $name;
     public string $description;
-    public int $price;
+    public float $price;
     public string $posted_At;
     public string $status;
     public string $image;
